Input: xpad - clarify LED enumeration
[ Upstream commit 1f6f02b728eef0d937d5138ae60dbf8fbd045f3e ]

Rename led_no -> pad_nr: the number stored there is not the LED number - it
gets translated later on to a LED number in xpad_identify_controller;

+++ drivers/input/joystick/xpad.c
@@ -343,7 +343,7 @@ struct usb_xpad {
 
 	int mapping;			/* map d-pad to buttons or to axes */
 	int xtype;			/* type of xbox device */
-	int pad_nr;			/* the order x360 pads were attached */
+	unsigned long pad_nr;			/* the order x360 pads were attached */
 	const char *name;		/* name of the device */
 };
